KS-2-1 integer variable

+++ src/main/kotlin/lesson_2/lesson2_task1.kt
@@ -5,8 +5,8 @@ fun main() {
     val mark2 = 4
     val mark3 = 3
     val mark4 = 5
-    val numberOfStudents = 4.0f
+    val numberOfStudents = 4
 
-    val average = (mark1 + mark2 + mark3 + mark4) / numberOfStudents
+    val average = (mark1 + mark2 + mark3 + mark4) / numberOfStudents.toFloat()
     println(average)
 }
